requirements.txt
flask==3.1.0
tensorflow-cpu==2.15.0
pandas==2.2.3
scikit-learn==1.6.1
requests==2.32.3
numpy==1.26.4